Changed implementation to use doubles instead of floats, sacrificing 0-30% performance for accuracy

+++ Data/TSPData.swift
@@ -32,7 +32,7 @@ extension TSPData {
         let lines = text.split(separator: "\n")
         let values = lines.map { $0.split(separator: ",") }
         let result = values.reduce(into: [Point]()) { result, next in
-            let point = flipped ? Point(Float(next[2])!, Float(next[1])!) : Point(Float(next[1])!, Float(next[2])!)
+            let point = flipped ? Point(Double(next[2])!, Double(next[1])!) : Point(Double(next[1])!, Double(next[2])!)
             result.append(point)
         }
         return result
